+++ packages/flate/lib/src/core.dart
@@ -6,21 +6,85 @@ import 'package:flutter/widgets.dart';
 import 'package:synchronized/synchronized.dart';
 import 'package:meta/meta.dart';
 
-part 'core/activity.dart';
-part 'core/fragment.dart';
-part 'core/fragments_registry.dart';
-part 'core/part.dart';
-part 'core/parts_registry.dart';
-part 'core/service.dart';
-part 'core/services_registry.dart';
 part 'core/store.dart';
-part 'core/types.dart';
 part 'core/registration.dart';
 part 'core/exceptions.dart';
 part 'core/element.dart';
-part 'core/context.dart';
-part 'core/context_registry.dart';
-part 'core/action.dart';
 part 'core/commit.dart';
-part 'core/resolvers.dart';
 part 'core/app_observer.dart';
+
+/// The class is used to store, initialize and dispose information about current Environment
+///
+/// Potentially, this class can contain information about the current configuration of the application,
+/// instances of global services necessary for the operation of services and other entities necessary
+/// for the correct operation of the application.
+abstract class FlateContext with FlateElementMixin {}
+
+/// Default instance of [FlateContext] used when context instance is not provided.
+class DefaultContext extends FlateContext {}
+
+/// This mixin should be used in declaring classes that interact with external resources
+/// such as a database, files or file system, etc.
+///
+/// Example:
+/// ```dart
+/// class MyLocationQuery extends FlateService {
+///   late final HttpClient _client;
+///
+///   @override
+///   FutureOr<void> initialize() async {
+///     super.initialize();
+///     _client = HttpClient();
+///   }
+///
+///   @override
+///   FutureOr<void> dispose() async {
+///     _client.close();
+///     super.dispose();
+///   }
+///
+///   Future<Map<String, dynamic>> myIpLocation() async {
+///     final uri = Uri.parse('http://ip-api.com/json/');
+///     final req = await _client.getUrl(uri);
+///     final res = await req.close();
+///     final data = await res.transform(utf8.decoder).join();
+///
+///     return jsonDecode(data);
+///   }
+/// }
+/// ```
+///
+abstract class FlateService with FlateElementMixin, FlateServiceMixin {}
+
+/// This mixin should be used in the declaration of classes that contain a shared data set
+/// and can be used by several [FlatePart] to ensure consistency at the level of several fragments.
+///
+/// Example:
+/// ```dart
+/// class User {}
+///
+/// class GuestUser implements User {}
+///
+/// class UserPart extends ChangeNotifier with SharedPart {
+///   late User _currentUser = GuestUser();
+///
+///   User get currentUser => _currentUser;
+///
+///   void changeCurrentUser(User newUser) {
+///     if (_currentUser == newUser) {
+///       return;
+///     }
+///     _currentUser = newUser;
+///     notifyListeners();
+///   }
+/// }
+///
+abstract class FlatePart with FlateElementMixin, FlatePartMixin {}
+
+/// The class is a fragment of the application state.
+///
+/// In cases when some part of the application state is not shared, then it should be placed in a fragment.
+/// The application can consist of one or more fragments. Each fragment in its implementation can use
+/// all registered parts and services, but must not use other fragments. Parts and services can be obtained
+/// in a fragment by calling methods [usePart] and [useService].
+abstract class FlateFragment with FlateElementMixin, FlateFragmentMixin {}