+++ packages/flate/lib/src/core/element.dart
@@ -1,6 +1,6 @@
 part of '../core.dart';
 
-abstract class FlateElement {
+mixin FlateElementMixin {
   static final _storeMount = Expando<FlateStore>('storeMount');
 
   FlateStore? get _store => _storeMount[this];
@@ -9,7 +9,7 @@ abstract class FlateElement {
   @protected
   bool get isMounted => _store != null;
 
-  /// This method used during registration to define set of [Type] with which this [FlateElement] will be registered
+  /// This method used during registration to define set of [Type] with which this [FlateElementMixin] will be registered
   ///
   /// By default this method provide type from [runtimeType] getter. This method can be overridden
   /// if an instance needs to be registered with multiple types that it implements.
@@ -46,8 +46,83 @@ abstract class FlateElement {
   @protected
   FutureOr<void> deactivate() {}
 
-  /// Mount provided [store] in this [FlateElement]
+  /// Mount provided [store] in this [FlateElementMixin]
   void _mount(FlateStore store) {
     _storeMount[this] = store;
   }
+
+  /// Provide instance of [FlateContext] by type [C]
+  ///
+  /// If instance for type [C] is not registered then [StateError] will be thrown.
+  /// If element is not mounted([isMounted] returns false) then [StateError] will be thrown
+  @protected
+  C _useContext<C>() {
+    if (isMounted) {
+      return _store!.useContext<C>();
+    }
+    throw StateError('Element of type $runtimeType is not mounted');
+  }
+
+  /// Provide instance of [FlatePart] by type [P]
+  ///
+  /// If instance for type [P] is not registered then [StateError] will be thrown.
+  /// If element is not mounted([isMounted] returns false) then [StateError] will be thrown
+  @protected
+  P _usePart<P>() {
+    if (isMounted) {
+      return _store!._usePart<P>();
+    }
+    throw StateError('Element is not mounted');
+  }
+
+  /// Provide instance of [FlateService] by type [S]
+  ///
+  /// If instance for type [S] is not registered then [StateError] will be thrown.
+  /// If element is not mounted([isMounted] returns false) then [StateError] will be thrown
+  @protected
+  S _useService<S>() {
+    if (isMounted) {
+      return _store!._useService<S>();
+    }
+    throw StateError('Element of type $runtimeType is not mounted');
+  }
+}
+
+/// Mixin used to implement [FlateService] class
+///
+/// Custom service type can be implemented by mixing [FlateElementMixin] and [FlateServiceMixin]
+mixin FlateServiceMixin on FlateElementMixin {
+  /// Provide instance of [FlateContext] by type [C]
+  @protected
+  C useContext<C>() => _useContext<C>();
+}
+
+/// Mixin used to implement [FlatePart] class
+///
+/// Custom part type can be implemented by mixing [FlateElementMixin] and [FlatePartMixin]
+mixin FlatePartMixin on FlateElementMixin {
+  /// Provide instance of [FlateContext] by type [C]
+  @protected
+  C useContext<C>() => _useContext<C>();
+
+  /// Provide instance of [FlateService] by type [S]
+  @protected
+  S useService<S>() => _useService<S>();
+}
+
+/// Mixin used to implement [FlateFragment] class
+///
+/// Custom part type can be implemented by mixing [FlateElementMixin] and [FlateFragmentMixin]
+mixin FlateFragmentMixin on FlateElementMixin {
+  /// Provide instance of [FlateContext] by type [C]
+  @protected
+  C useContext<C>() => _useContext<C>();
+
+  /// Provide instance of [FlateService] by type [S]
+  @protected
+  S useService<S>() => _useService<S>();
+
+  /// Provide instance of [FlatePart] by type [P]
+  @protected
+  P usePart<P>() => _usePart<P>();
 }