+++ packages/flate/test/app_observer_test.dart
@@ -1,0 +1,2 @@
+// TODO(bysensa): write tests for AppObserver API
+void main() {}