+++ packages/flate/CHANGELOG.md
@@ -9,3 +9,7 @@
 ## 0.0.1+2
 
 * Implement AppObserver API to allow elements handle memory pressure and app lifecycle change
+* Change code organization
+* Support creating custom variations of `FlateService`, `FlatePart`, `FlateFragment`
+* Implement `FlateComponentMixin` for usage with `StatefulWidget`
+* Remove `onNotification` method from FlateComponentModel